pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "tea"
version = "0.4.0"
description = "Manage tmuxp session configurations"
authors = ["Kiyoshi '13k' Murata <kbmurata@gmail.com>"]

[tool.poetry.scripts]
tea = 'tea.cli:main'

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
tmuxp = "^1.9.2"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
isort = "^5.8.0"
mypy = "^0.812"
ptpython = "^3.0.17"
pylint = "^2.8.3"
rope = "^0.19.0"

[tool.black]
line-length = 98
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | tmp
  )/
)
'''

[tool.isort]
profile = "black"

[tool.pylint.messages_control]
disable = "C0330, C0326"

[tool.pylint.format]
max-line-length = 98
